Add files via upload
After cleaning the data using Excel, the data was imported into Power BI to create various visualizations, including charts and graphs. The goal of this project was to gain a better understanding of the average salary across different fields in marketing, as well as the average salary for entry-level positions in the industry. By creating these visualizations, the data was presented in a clear and concise manner, making it easier to interpret and draw insights from.